Move PriceLookup and EstimatedPriceCalculator to `logistics` for convenience
