+++ logistics/src/main/java/com/alevel/deliverit/gateway/PriceLookup.java
@@ -1,12 +1,13 @@
 package com.alevel.deliverit.gateway;
 
 import com.alevel.deliverit.BusinessLogicService;
-import com.alevel.deliverit.EstimatedPriceCalculator;
 import com.alevel.deliverit.Subscribe;
 import com.alevel.deliverit.billing.Money;
 import com.alevel.deliverit.customers.request.PriceLookupRequest;
 
 import static com.alevel.deliverit.SubscribeAddress.BILLING_CALCULATE_PRICE;
+
+//import static com.alevel.deliverit.SubscribeAddress.BILLING_CALCULATE_PRICE;
 
 /**
  * @author Sergey Bogovesov
@@ -15,6 +16,7 @@ public class PriceLookup implements BusinessLogicService {
 
     @Subscribe(BILLING_CALCULATE_PRICE)
     public static Money calculate(PriceLookupRequest request) {
+        System.out.println(request);
         EstimatedPriceCalculator estimatedPriceCalculator = new EstimatedPriceCalculator();
         return estimatedPriceCalculator.calculate(request.getParcelWeight(), request.getRoute());
     }
